Add Account Assignment test for if a policy is missing
When a permission set is assigned with a customer-managed policy that
doesn't exist, AWS returns the error message in the body of the request.
We want to make sure this is exposed to the user so they can see which
policy is missing. Previously this returned the error message from
the connection which would just be `nil` if the
`DescribeAccountAssignmentCreationStatus` request is successful.

Without knowing a failure mode for this it'd be hard to create a test
for it. However the type returned is the same as the account creation
method so the logic would be the same.